+++ Assets/_Project/Scripts/Actors/Enemy/EnemyStats.cs
@@ -9,11 +9,11 @@ public class EnemyStats
 
     [Header("Properties")]
     public Affinity Affinity;
-    public int MaxHp;
-    public int CurrentHp;
+    public float MaxHp;
+    public float CurrentHp;
     public float MovementSpeed;
-    public int Attack;
+    public float Attack;
     public float AttackSpeed;
-    public int Defense;
+    public float Defense;
     public float AttackRange;
 }
